+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 0.50 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added `content.get_transitions` to be able to get available transition ids
+  (as `wfTool.getTransitionsFor` returns a list of dict with transition infos).
+  [gbastien]
 
 0.49 (2021-11-08)
 -----------------